fix: 修复元素位置计算问题
